+++ src/routes/ViewJobLog.js
@@ -4,13 +4,17 @@ import { FormattedMessage } from 'react-intl';
 import SafeHTMLMessage from '@folio/react-intl-safe-html';
 
 import { stripesConnect } from '@folio/stripes/core';
+import { Headline } from '@folio/stripes/components';
 import { Preloader } from '@folio/stripes-data-transfer-components';
 
 import { LogViewer } from '../components/LogViewer';
 import { LANGUAGES } from '../components/CodeHighlight/Languages';
 import { THEMES } from '../components';
 
-import { LOG_VIEWER } from '../utils';
+import {
+  DATA_TYPES,
+  LOG_VIEWER,
+} from '../utils';
 
 import sharedCss from '../shared.css';
 
@@ -21,8 +25,12 @@ export class ViewJobLog extends Component {
   static manifest = Object.freeze({
     jobLog: {
       type: 'okapi',
-      path: 'metadata-provider/jobLogEntries/:{id}/records/:{recordId}',
       throwsErrors: false,
+      path: (_q, _p) => {
+        const recordId = _q.instanceLineId || _p.recordId;
+
+        return `metadata-provider/jobLogEntries/${_p.id}/records/${recordId}`;
+      },
     },
     srsMarcBib: {
       type: 'okapi',
@@ -44,6 +52,18 @@ export class ViewJobLog extends Component {
     items: {
       type: 'okapi',
       path: 'inventory/items',
+      throwsErrors: false,
+      accumulate: true,
+    },
+    invoice: {
+      type: 'okapi',
+      path: 'invoice-storage/invoices',
+      throwsErrors: false,
+      accumulate: true,
+    },
+    invoiceLine: {
+      type: 'okapi',
+      path: 'invoice-storage/invoice-lines',
       throwsErrors: false,
       accumulate: true,
     },
@@ -71,11 +91,21 @@ export class ViewJobLog extends Component {
         hasLoaded: PropTypes.bool.isRequired,
         records: PropTypes.arrayOf(PropTypes.object.isRequired).isRequired,
       }),
+      invoice: PropTypes.shape({
+        hasLoaded: PropTypes.bool.isRequired,
+        records: PropTypes.arrayOf(PropTypes.object.isRequired).isRequired,
+      }),
+      invoiceLine: PropTypes.shape({
+        hasLoaded: PropTypes.bool.isRequired,
+        records: PropTypes.arrayOf(PropTypes.object.isRequired).isRequired,
+      }),
     }).isRequired,
     mutator: PropTypes.shape({
       instances: PropTypes.shape({ GET: PropTypes.func.isRequired }).isRequired,
       holdings: PropTypes.shape({ GET: PropTypes.func.isRequired }).isRequired,
       items: PropTypes.shape({ GET: PropTypes.func.isRequired }).isRequired,
+      invoice: PropTypes.shape({ GET: PropTypes.func.isRequired }).isRequired,
+      invoiceLine: PropTypes.shape({ GET: PropTypes.func.isRequired }).isRequired,
     }).isRequired,
   };
 
@@ -92,6 +122,8 @@ export class ViewJobLog extends Component {
       this.fetchInstancesData();
       this.fetchHoldingsData();
       this.fetchItemsData();
+      this.fetchInvoiceData();
+      this.fetchInvoiceLineData();
     }
   }
 
@@ -119,14 +151,35 @@ export class ViewJobLog extends Component {
     });
   }
 
+  fetchInvoiceData() {
+    const invoiceIds = this.props.resources.jobLog.records[0]?.relatedInvoiceInfo.idList || [];
+
+    invoiceIds.forEach(invoiceId => {
+      if (invoiceId) {
+        this.props.mutator.invoice.GET({ path: `invoice-storage/invoices/${invoiceId}` });
+      }
+    });
+  }
+
+  fetchInvoiceLineData() {
+    const invoiceLineId = this.props.resources.jobLog.records[0]?.relatedInvoiceLineInfo.id;
+
+    if (invoiceLineId) {
+      this.props.mutator.invoiceLine.GET({ path: `invoice-storage/invoice-lines/${invoiceLineId}` });
+    }
+  }
+
   get jobLogData() {
     const { resources } = this.props;
 
     const jobLog = resources.jobLog || {};
+    const srsMarcBibLog = resources.srsMarcBib || {};
     const [record] = jobLog.records || [];
+    const recordType = srsMarcBibLog.records[0]?.recordType;
 
     return {
       hasLoaded: jobLog.hasLoaded,
+      recordType,
       record,
     };
   }
@@ -179,6 +232,30 @@ export class ViewJobLog extends Component {
     };
   }
 
+  get invoiceData() {
+    const { resources } = this.props;
+
+    const invoice = resources.invoice || {};
+    const [record] = invoice.records || [];
+
+    return {
+      hasLoaded: invoice.hasLoaded,
+      record,
+    };
+  }
+
+  get invoiceLineData() {
+    const { resources } = this.props;
+
+    const invoiceLine = resources.invoiceLine || {};
+    const [record] = invoiceLine.records || [];
+
+    return {
+      hasLoaded: invoiceLine.hasLoaded,
+      record,
+    };
+  }
+
   getErrorMessage(entityOption) {
     const {
       hasLoaded,
@@ -194,6 +271,7 @@ export class ViewJobLog extends Component {
       relatedItemInfo,
       relatedOrderInfo,
       relatedInvoiceInfo,
+      relatedInvoiceLineInfo,
     } = record;
 
     const errors = {
@@ -202,7 +280,10 @@ export class ViewJobLog extends Component {
       [OPTIONS.HOLDINGS]: relatedHoldingsInfo.error,
       [OPTIONS.ITEM]: relatedItemInfo.error,
       [OPTIONS.ORDER]: relatedOrderInfo.error,
-      [OPTIONS.INVOICE]: relatedInvoiceInfo.error,
+      [OPTIONS.INVOICE]: {
+        invoiceInfo: relatedInvoiceInfo.error,
+        invoiceLineInfo: relatedInvoiceLineInfo.error,
+      },
     };
 
     return errors[entityOption];
@@ -211,6 +292,7 @@ export class ViewJobLog extends Component {
   render() {
     const {
       hasLoaded,
+      recordType,
       record,
     } = this.jobLogData;
 
@@ -227,8 +309,10 @@ export class ViewJobLog extends Component {
     const {
       sourceRecordOrder,
       sourceRecordTitle,
+      relatedInvoiceLineInfo: { fullInvoiceLineNumber },
     } = record;
 
+    const isEdifactType = recordType === DATA_TYPES[1];
     const toolbar = {
       visible: true,
       message: (
@@ -236,21 +320,60 @@ export class ViewJobLog extends Component {
           id="ui-data-import.import-log"
           tagName="span"
           values={{
-            recordOrder: sourceRecordOrder + 1,
+            recordOrder: isEdifactType ? fullInvoiceLineNumber : sourceRecordOrder + 1,
             recordTitle: sourceRecordTitle,
           }}
         />
       ),
       showThemes: false,
+      activeFilter: isEdifactType ? OPTIONS.INVOICE : OPTIONS.SRS_MARC_BIB,
     };
 
     const logs = {
-      [OPTIONS.SRS_MARC_BIB]: this.srsMarcBibData.record,
-      [OPTIONS.INSTANCE]: this.instanceData.record,
-      [OPTIONS.HOLDINGS]: this.holdingsData.record,
-      [OPTIONS.ITEM]: this.itemData.record,
-      [OPTIONS.ORDER]: {},
-      [OPTIONS.INVOICE]: {},
+      [OPTIONS.SRS_MARC_BIB]: [{
+        label: '',
+        logs: this.srsMarcBibData.record,
+        error: this.getErrorMessage(OPTIONS.SRS_MARC_BIB),
+        errorBlockId: 'srs-marc-bib-error',
+      }],
+      [OPTIONS.INSTANCE]: [{
+        label: '',
+        logs: this.instanceData.record,
+        error: this.getErrorMessage(OPTIONS.INSTANCE),
+        errorBlockId: 'instance-error',
+      }],
+      [OPTIONS.HOLDINGS]: [{
+        label: '',
+        logs: this.holdingsData.record,
+        error: this.getErrorMessage(OPTIONS.HOLDINGS),
+        errorBlockId: 'holdings-error',
+      }],
+      [OPTIONS.ITEM]: [{
+        label: '',
+        logs: this.itemData.record,
+        error: this.getErrorMessage(OPTIONS.ITEM),
+        errorBlockId: 'item-error',
+      }],
+      [OPTIONS.ORDER]: [{}],
+      [OPTIONS.INVOICE]: [{
+        label: (
+          <Headline margin="none">
+            <FormattedMessage id="ui-data-import.logViewer.invoiceLine" />
+          </Headline>
+        ),
+        logs: this.invoiceLineData.record,
+        error: this.getErrorMessage(OPTIONS.INVOICE).invoiceLineInfo,
+        errorBlockId: 'invoice-line-error',
+      }, {
+        label: (
+          <Headline margin="none">
+            <FormattedMessage id="ui-data-import.logViewer.invoice" />
+          </Headline>
+        ),
+        logs: this.invoiceData.record,
+        error: this.getErrorMessage(OPTIONS.INVOICE).invoiceInfo,
+        errorBlockId: 'invoice-error',
+      }],
     };
 
     return (
@@ -260,7 +383,6 @@ export class ViewJobLog extends Component {
           language={LANGUAGES.JSON}
           theme={THEMES.COY}
           toolbar={toolbar}
-          errorDetector={entityOption => this.getErrorMessage(entityOption)}
         />
       </div>
     );